# src/main/resources/static/openapi.yaml
openapi: 3.1.0
info:
  title: Sweet & Sweat Backend API
  version: v0.1.0
  description: API contract for the midterm demo flow.
servers:
  - url: http://100.89.171.113:8080
    description: Tailscale development server
  - url: http://localhost:8080
    description: Local development server
...
security:
  - bearerAuth: []
paths:
...
  /api/auth/signup:
    post:
      tags: [Auth]
      summary: Create a demo user account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupRequest"
      responses:
        "201":
          description: User account created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseUserProfileResponse"
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login and issue JWT tokens
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login succeeded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseLoginResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    ApiResponseMeta:
      type: object
      required: [success, code, message, timestamp]
      properties:
        success:
          type: boolean
          example: true
        code:
          type: string
          example: OK
        message:
          type: string
          example: Request succeeded
        timestamp:
          type: string
          format: date-time
...
    ApiResponseUserProfileResponse:
      allOf:
        - $ref: "#/components/schemas/ApiResponseMeta"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/UserProfileResponse"
    ApiResponseLoginResponse:
      allOf:
        - $ref: "#/components/schemas/ApiResponseMeta"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/LoginResponse"
...
    SignupRequest:
      type: object
      required: [loginId, password, nickname]
      properties:
        loginId:
          type: string
          example: demo
        password:
          type: string
          format: password
          example: demo1234
        nickname:
          type: string
          example: Demo User
    LoginRequest:
      type: object
      required: [loginId, password]
      properties:
        loginId:
          type: string
          example: demo
        password:
          type: string
          format: password
          example: demo1234
    LoginResponse:
      type: object
      required: [accessToken, refreshToken, tokenType, user]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        tokenType:
          type: string
          example: Bearer
        user:
          $ref: "#/components/schemas/UserProfileResponse"
...
    UserProfileResponse:
      type: object
      required: [id, loginId, nickname]
      properties:
        id:
          type: integer
          format: int64
        loginId:
          type: string
        nickname:
          type: string
        exerciseExperience:
          type: string
          nullable: true
        preferredExercise:
          type: string
          nullable: true
...